provider/libserver: size_t correctness for GetObjectSize

+++ provider/libserver/ECGenericObjectTable.h
@@ -71,7 +71,7 @@ public:
     ECRESULT UpdateMinMax(const sObjectTableKey &sKey, unsigned int i, struct propVal *lpPropVal, bool fMax, bool *lpfModified);
     ECRESULT UpdateMinMaxRemove(const sObjectTableKey &sKey, unsigned int i, bool fMax, bool *lpfModified);
 	
-	unsigned int GetObjectSize(void) const;
+	size_t GetObjectSize(void) const;
 
     struct propVal *m_lpProps;
     unsigned int m_cProps;
@@ -184,7 +184,7 @@ public:
 	virtual	ECRESULT	GetPropCategory(struct soap *soap, unsigned int ulPropTag, sObjectTableKey sKey, struct propVal *lpPropVal);
 	virtual unsigned int GetCategories();
 
-	virtual unsigned int GetObjectSize(void);
+	virtual size_t GetObjectSize(void);
 
 protected:
 	// Add an actual row to the table, and send a notification if required. If you add an existing row, the row is modified and the notification is send as a modification
